Partially-sorted observations
A two-element sorting function is clearly commutative.

Performing two identical observations and then sorting the results is
also clearly commutative.

Less obviously, the function is also commutative if only part of the
result is used for sorting, provided that the unsorted part is only
examined when the compared parts differ. This can be useful when part of
the observation has an uncomparable type, a function for example.