observe_eq, observe_eq_and
It was hard to remember the type of observe_eq, because its name makes
it look as if only the equality is observed. Since in fact it is both
equality and possibly a second arbitrary observation which is observed,
that function has been renamed to observe_eq_and.

A new function named observe_eq, based on observe_eq_and, has been added
to implement the simpler, equality-only observation.