Data.Unordered
Instead of using primitives, it is much more convenient to make one or
more arbitrary observations on each argument. For this to work, it must
be possible to hide which observation comes from which argument; those
types for which this is possible are called "Unorderable", because it is
possible to remove ordering information from a pair of observations.